[#2502] moving code around for resource coverage update
