+++ hs_composite_resource/tests/test_composite_resource.py
@@ -9,7 +9,6 @@ from hs_core import hydroshare
 from hs_core.models import BaseResource
 from hs_core.hydroshare.utils import resource_file_add_process
 from hs_core.views.utils import create_folder, move_or_rename_file_or_folder, remove_folder
-from hs_file_types.utils import update_resource_temporal_coverage, update_resource_spatial_coverage
 
 from hs_file_types.models import GenericLogicalFile, GeoRasterLogicalFile, GenericFileMetaData
 from hs_file_types.tests.utils import CompositeResourceTestMixin
@@ -500,7 +499,7 @@ class CompositeResourceTest(MockIRODSTestCaseMixin, TransactionTestCase,
 
         # test updating the resource coverage by user action - which should update the resource
         # coverage as a superset of all coverages of all the contained aggregations/logical files
-        update_resource_temporal_coverage(self.composite_resource)
+        self.composite_resource.update_temporal_coverage()
         res_coverage = self.composite_resource.metadata.coverages.all().filter(
             type='period').first()
         raster_lfo_coverage = raster_logical_file.metadata.coverages.all().filter(
@@ -538,8 +537,8 @@ class CompositeResourceTest(MockIRODSTestCaseMixin, TransactionTestCase,
 
         # test updating the resource coverage by user action - which should update the resource
         # coverage as a superset of all coverages of all the contained aggregations/logical files
-        update_resource_temporal_coverage(self.composite_resource)
-        update_resource_spatial_coverage(self.composite_resource)
+        self.composite_resource.update_temporal_coverage()
+        self.composite_resource.update_spatial_coverage()
         # resource temporal coverage is now super set of the 2 temporal coverages
         # in 2 LFOs
         res_coverage = self.composite_resource.metadata.coverages.all().filter(
@@ -559,7 +558,7 @@ class CompositeResourceTest(MockIRODSTestCaseMixin, TransactionTestCase,
         generic_logical_file.metadata.create_element('coverage', type='point', value=value_dict)
 
         # update resource spatial coverage from aggregations spatial coverages
-        update_resource_spatial_coverage(self.composite_resource)
+        self.composite_resource.update_spatial_coverage()
         res_coverage = self.composite_resource.metadata.coverages.all().filter(
             type='box').first()
         self.assertEqual(res_coverage.value['projection'], 'WGS 84 EPSG:4326')
@@ -577,7 +576,7 @@ class CompositeResourceTest(MockIRODSTestCaseMixin, TransactionTestCase,
                                                      type='box', value=value_dict)
 
         # update resource spatial coverage from aggregations spatial coverages
-        update_resource_spatial_coverage(self.composite_resource)
+        self.composite_resource.update_spatial_coverage()
 
         res_coverage = self.composite_resource.metadata.coverages.all().filter(
             type='box').first()
